+++ gems/ast-merge/lib/ast/merge.rb
@@ -89,6 +89,7 @@ module Ast
     autoload :SectionTyping, 'ast/merge/section_typing'
     autoload :SmartMergerBase, 'ast/merge/smart_merger_base'
     autoload :SourceRegionReportSupport, 'ast/merge/source_region_report_support'
+    autoload :SourceRender, 'ast/merge/source_render'
     autoload :StructuralEdit, 'ast/merge/structural_edit'
     autoload :StructuredEmitterProvenanceSupport, 'ast/merge/structured_emitter_provenance_support'
     autoload :StructuredReviewApplySupport, 'ast/merge/structured_review_apply_support'